Fixed last bugs

+++ src/main/java/com/neon/libary/interfaces/ProjectileEntity.java
@@ -10,5 +10,5 @@ package com.neon.libary.interfaces;
  * @author Mathias
  */
 public interface ProjectileEntity extends Entity {
-    int getDamage();
+    float getDamage();
 }
